api,bus,worker: cleanup clients and api
