+++ bus/client/sectors.go
@@ -7,6 +7,7 @@ import (
 	"go.sia.tech/core/types"
 )
 
-func (c *Client) DeleteHostSector(ctx context.Context, hk types.PublicKey, root types.Hash256) error {
-	return c.c.WithContext(ctx).DELETE(fmt.Sprintf("/sectors/%s/%s", hk, root))
+// DeleteHostSector deletes the given sector on host with given host key.
+func (c *Client) DeleteHostSector(ctx context.Context, hostKey types.PublicKey, sectorRoot types.Hash256) error {
+	return c.c.WithContext(ctx).DELETE(fmt.Sprintf("/sectors/%s/%s", hostKey, sectorRoot))
 }
